requirements.txt
Babel==2.6.0
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.13
django-debug-toolbar==1.11
django-extra-views==0.11.0
django-haystack==2.8.1
django-localflavor==2.1
django-oscar==1.6.7
django-oscar-accounts==1.0.0
django-oscar-paypal==1.0.0
django-phonenumber-field==2.0.1
django-tables2==1.21.2
django-treebeard==4.3
django-widget-tweaks==1.4.3
environ==1.0
factory-boy==2.11.1
Faker==1.0.5
gunicorn==19.9.0
idna==2.8
mock==2.0.0
pbr==5.2.0
phonenumbers==8.10.10
Pillow==6.0.0
psycopg2-binary==2.8.2
purl==1.5
pycountry==18.12.8
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
six==1.12.0
sorl-thumbnail==12.4.1
sqlparse==0.3.0
text-unidecode==1.2
Unidecode==1.0.23
urllib3==1.24.2
whitenoise==4.1.2
wincertstore==0.2
